[17] Use common_cli in semantic extraction and clustering.
